requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
boto3==1.26.145
    # via pybites-tools
botocore==1.29.145
    # via
    #   boto3
    #   s3transfer
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
idna==3.4
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
pillow==9.5.0
    # via -r requirements.in
pybites-tools==0.2
    # via -r requirements.in
pyperclip==1.8.2
    # via pybites-tools
python-dateutil==2.8.2
    # via
    #   botocore
    #   pybites-tools
python-dotenv==1.0.0
    # via pybites-tools
requests==2.31.0
    # via
    #   -r requirements.in
    #   pybites-tools
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.16
    # via
    #   botocore
    #   requests
